+++ h2integrate/core/supported_models.py
@@ -1,4 +1,5 @@
 from h2integrate.resource.river import RiverResource
+from h2integrate.core.feedstocks import FeedstockCostModel, FeedstockPerformanceModel
 from h2integrate.transporters.pipe import PipePerformanceModel
 from h2integrate.transporters.cable import CablePerformanceModel
 from h2integrate.converters.steel.steel import SteelPerformanceModel, SteelCostAndFinancialModel
@@ -50,6 +51,10 @@ from h2integrate.converters.methanol.co2h_methanol_plant import (
     CO2HMethanolPlantPerformanceModel,
 )
 from h2integrate.control.control_rules.storage.h2_storage import PyomoDispatchH2Storage
+from h2integrate.converters.natural_gas.natural_gas_cc_ct import (
+    NaturalGasCostModel,
+    NaturalGasPerformanceModel,
+)
 from h2integrate.converters.hydrogen.singlitico_cost_model import SingliticoCostModel
 from h2integrate.converters.co2.marine.direct_ocean_capture import DOCCostModel, DOCPerformanceModel
 from h2integrate.control.control_strategies.pyomo_controllers import (
@@ -78,6 +83,9 @@ from h2integrate.converters.co2.marine.ocean_alkalinity_enhancement import (
     OAEPerformanceModel,
     OAECostAndFinancialModel,
 )
+from h2integrate.converters.hydrogen.custom_electrolyzer_cost_model import (
+    CustomElectrolyzerCostModel,
+)
 from h2integrate.converters.hydrogen.geologic.stimulated_geoh2_plant import (
     StimulatedGeoH2CostModel,
     StimulatedGeoH2FinanceModel,
@@ -102,10 +110,11 @@ supported_models = {
     "eco_pem_electrolyzer_performance": ECOElectrolyzerPerformanceModel,
     "singlitico_electrolyzer_cost": SingliticoCostModel,
     "basic_electrolyzer_cost": BasicElectrolyzerCostModel,
+    "custom_electrolyzer_cost": CustomElectrolyzerCostModel,
+    "wombat": WOMBATElectrolyzerModel,
     "simple_ASU_cost": SimpleASUCostModel,
     "simple_ASU_performance": SimpleASUPerformanceModel,
     "hopp": HOPPComponent,
-    "wombat": WOMBATElectrolyzerModel,
     "reverse_osmosis_desalination_performance": ReverseOsmosisPerformanceModel,
     "reverse_osmosis_desalination_cost": ReverseOsmosisCostModel,
     "simple_ammonia_performance": SimpleAmmoniaPerformanceModel,
@@ -131,6 +140,8 @@ supported_models = {
     "stimulated_geoh2_performance": StimulatedGeoH2PerformanceModel,
     "stimulated_geoh2_cost": StimulatedGeoH2CostModel,
     "stimulated_geoh2": StimulatedGeoH2FinanceModel,
+    "natural_gas_performance": NaturalGasPerformanceModel,
+    "natural_gas_cost": NaturalGasCostModel,
     # Transport
     "cable": CablePerformanceModel,
     "pipe": PipePerformanceModel,
@@ -152,8 +163,11 @@ supported_models = {
     "pyomo_dispatch_electrolyzer": PyomoDispatchElectrolyzer,
     "pyomo_dispatch_h2_storage": PyomoDispatchH2Storage,
     "pyomo_dispatch_battery": PyomoDispatchBattery,
+    # Feedstock
+    "feedstock_performance": FeedstockPerformanceModel,
+    "feedstock_cost": FeedstockCostModel,
     # Finance
     "ProFastComp": ProFastComp,
 }
 
-electricity_producing_techs = ["wind", "solar", "pv", "river", "hopp"]
+electricity_producing_techs = ["wind", "solar", "pv", "river", "hopp", "natural_gas_plant"]